move components into separate folders
