+++ src/components/Header/Header.module.css
@@ -8,7 +8,6 @@
 
 .NavItem {
   display: block;
-  height: 100%;
   margin-right: 1em;
 }
