+++ src/hoc/withError.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import Modal from "../components/Modal";
+import Modal from "../components/Modal/Modal";
 
 const withError = WrappedComponent => {
   return function(props) {
